+++ libp2p/src/builder/phase/other_transport.rs
@@ -9,6 +9,7 @@ use libp2p_core::{Negotiated, UpgradeInfo};
 #[cfg(feature = "relay")]
 use libp2p_identity::PeerId;
 
+#[allow(deprecated)]
 use crate::bandwidth::BandwidthSinks;
 use crate::SwarmBuilder;
 
@@ -176,20 +177,42 @@ impl<T: AuthenticatedMultiplexedTransport, Provider>
 impl<Provider, T: AuthenticatedMultiplexedTransport>
     SwarmBuilder<Provider, OtherTransportPhase<T>>
 {
+    #[allow(deprecated)]
+    #[deprecated(note = "Use `with_bandwidth_metrics` instead.")]
     pub fn with_bandwidth_logging(
         self,
     ) -> (
         SwarmBuilder<
             Provider,
-            BehaviourPhase<impl AuthenticatedMultiplexedTransport, NoRelayBehaviour>,
+            BandwidthMetricsPhase<impl AuthenticatedMultiplexedTransport, NoRelayBehaviour>,
         >,
         Arc<BandwidthSinks>,
     ) {
+        #[allow(deprecated)]
         self.without_any_other_transports()
             .without_dns()
             .without_websocket()
             .without_relay()
             .with_bandwidth_logging()
+    }
+}
+#[cfg(feature = "metrics")]
+impl<Provider, T: AuthenticatedMultiplexedTransport>
+    SwarmBuilder<Provider, OtherTransportPhase<T>>
+{
+    pub fn with_bandwidth_metrics(
+        self,
+        registry: &mut libp2p_metrics::Registry,
+    ) -> SwarmBuilder<
+        Provider,
+        BehaviourPhase<impl AuthenticatedMultiplexedTransport, NoRelayBehaviour>,
+    > {
+        self.without_any_other_transports()
+            .without_dns()
+            .without_websocket()
+            .without_relay()
+            .without_bandwidth_logging()
+            .with_bandwidth_metrics(registry)
     }
 }
 impl<Provider, T: AuthenticatedMultiplexedTransport>